separating logic
